requirements.txt
streamlit==1.32.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
pyarrow==15.0.2
scikit-learn==1.4.2
pillow==10.3.0
